This pull request solves 4 problems:

1. Migrates `Tensor.random_()` from TH to ATen, both CPU and CUDA versions.
2. Allows `random_()` to generate full range of 64 bit numbers(including unsigned 64 bit max value).
3. Implements `random_()` for boolean tensors on CUDA.
4. Makes `random_()` template methods to allow using it with custom RNGs.

 It is done by the following changes:

1. Drop TH CPU implementations of `random_()`.
2. Change API of `random_()` to make to argument optional to allow call `random_(from=min_value, to=None)` that generates full 64 bit range numbers.
3. Make three native functions `random_()`, `random_(to)` and `random_(to, from)` which call three different kernels(both CPU and CUDA).
4. Create three random_ kernels(both CPU and CUDA) to handle `random_(no params)`, `random(to, from)` and `random(from=min_value, to=None)` cases.
5. Templatize all random_ implementations and kernels to use them with custom RNGs.
6. Create C++ tests that uses custom RNG and `random_()` templates to check correctness.
7. Create Python tests to cover all `random_()` scenarios with all possible dtypes and devices.

Fixes #24752
Fixes #32510
Fixes #33299
Fixes #33725

Known issues:
#33793 random_() is not supported for bfloat16 CUDA tensors on Windows


Differential Revision: [D20056350](https://our.internmc.facebook.com/intern/diff/D20056350)

[ghstack-poisoned]

+++ aten/src/ATen/native/cuda/DistributionTemplates.h
@@ -170,7 +170,7 @@ template<typename RNG>
 void random_from_to_kernel(TensorIterator& iter, uint64_t range, int64_t base, RNG* gen) {
 #ifdef _WIN32
   // TODO: https://github.com/pytorch/pytorch/issues/33793
-  if (iter.dtype() == ScalarType::Half) {
+  if (iter.dtype() == ScalarType::BFloat16) {
     TORCH_CHECK(false, "random_() is not supported for bfloat16 CUDA tensors on Windows. Please see https://github.com/pytorch/pytorch/issues/33793");
   }
 #endif
@@ -216,7 +216,7 @@ template<typename RNG>
 void random_full_64_bits_range_kernel(TensorIterator& iter, RNG* gen) {
 #ifdef _WIN32
   // TODO: https://github.com/pytorch/pytorch/issues/33793
-  if (iter.dtype() == ScalarType::Half) {
+  if (iter.dtype() == ScalarType::BFloat16) {
     TORCH_CHECK(false, "random_() is not supported for bfloat16 CUDA tensors on Windows. Please see https://github.com/pytorch/pytorch/issues/33793");
   }
 #endif
@@ -258,7 +258,7 @@ template<typename RNG>
 void random_kernel(TensorIterator& iter, RNG* gen) {
 #ifdef _WIN32
   // TODO: https://github.com/pytorch/pytorch/issues/33793
-  if (iter.dtype() == ScalarType::Half) {
+  if (iter.dtype() == ScalarType::BFloat16) {
     TORCH_CHECK(false, "random_() is not supported for bfloat16 CUDA tensors on Windows. Please see https://github.com/pytorch/pytorch/issues/33793");
   }
 #endif
